requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
APScheduler==3.11.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
categories==0.3.8
certifi==2025.1.31
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.3
django-jazzmin==3.0.1
docx==0.2.4
fonttools==4.59.0
frozenlist==1.6.0
greenlet==3.2.0
idna==3.10
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.4.0
magic-filter==1.0.12
matplotlib==3.10.3
multidict==6.4.3
numpy==2.2.6
packaging==25.0
parts==2.1.0
pillow==11.2.1
products==2.0.1
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
SQLAlchemy==2.0.25
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
users==1.0.dev0
win32_setctime==1.2.0
yarl==1.20.0